DEV: Extract suite/test setup
